+++ libIntegra/src/file_helper.cpp
@@ -274,8 +274,8 @@ namespace integra_internal
 
 	string CFileHelper::ensure_filename_has_suffix( const string &filename, const string &suffix )
 	{
-		int filename_length = filename.length();
-		int suffix_length = suffix.length();
+		int filename_length = (int)(filename.length());
+		int suffix_length = (int)(suffix.length());
 
 		if( filename_length > suffix_length + 1 )
 		{
